pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fluxsim"
version = "0.1.0"
description = "FluxLab network simulator: DNS fast-flux, LB, and CDN labs in Docker"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Leslie Etubo" }]

# ⬇️ Added deps for agents + compose generation
dependencies = [
  "riposte>=0.3",
  "dnspython>=2.6,<3",
  "kafka-python>=2,<3",
  "requests>=2.31,<3",
  "PyYAML>=6.0"        # you import `yaml` in compose_gen.py
]

[project.optional-dependencies]
dev = [
  "black~=23.12",
  "ruff~=0.1",
  "mypy~=1.8",
  "pytest>=7.4,<9",
  "pytest-cov>=4.1,<5",
  "types-requests",
]

[project.scripts]
fluxsim = "fluxsim.cli:main"

[tool.setuptools]
include-package-data = true

# Tell setuptools to ONLY package the Python package
[tool.setuptools.packages.find]
where = ["."]
include = ["fluxsim*"]
exclude = ["tests*", "html*", "dns_config*", "nginx_configs*", "scripts*", ".github*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B"]
ignore = ["E203", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["fluxsim"]
combine-as-imports = true

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]
